requirements.txt
faiss_cpu==1.9.0.post1
langchain_community==0.3.13
langchain_core==0.3.28
langchain==0.3.13
numpy==1.26.4
pypdf==5.1.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.41.1
